Revert "feat: 회원 권한 관리 및 로그인 회원가입 기능 구현 완료"

+++ src/main/java/bdbe/bdbd/review/ReviewResponse.java
@@ -14,7 +14,7 @@ public class ReviewResponse {
         private Long cId;
         private Reservation reservation; //1:1, 참조용(read-only)
         private String comment;
-        private double rate;
+        private int rate;
 
         public getReviewById(Review review) {
             this.id = review.getId();
